move photometric series model into its own file
